Refactor Form and Painting
