+++ src/App.vue
@@ -3,43 +3,43 @@
 </template>
 
 <script>
-import Main from './components/Main.vue';
+import Main from "./components/Main.vue";
 
 export default {
-  name: 'App',
+  name: "App",
   metaInfo: {
-    title: 'Pelle Öhrn, copywriter',
+    title: "Pelle Öhrn, copywriter",
     meta: [
-      { keywords: 'Öhrn, copywriter, Pelle Öhrn, copywriter Karlstad, Berghs, konsult, korrekturläsning, korrekturläsning Karlstad' },
-      { charset: 'utf-8' },
+      { keywords: "Öhrn, copywriter, Pelle Öhrn, copywriter Karlstad, Berghs, konsult, korrekturläsning, korrekturläsning Karlstad" },
+      { charset: "utf-8" },
       {
-        property: 'og:title',
-        content: 'Pelle Öhrn, copywriter',
-        vmid: 'og:title',
+        property: "og:title",
+        content: "Pelle Öhrn, copywriter",
+        vmid: "og:title",
       },
       {
-        property: 'og:url',
-        content: 'https://www.pelleohrn.se/',
-        vmid: 'og:url',
+        property: "og:url",
+        content: "https://www.pelleohrn.se/",
+        vmid: "og:url",
       },
       {
-        property: 'og:image',
-        content: 'https://www.pelleohrn.se/images/pelle.jpg',
-        vmid: 'og:image',
+        property: "og:image",
+        content: "https://www.pelleohrn.se/images/pelle.jpg",
+        vmid: "og:image",
       },
       {
-        property: 'og:description',
-        content: 'Jag gillar att vränga ord, hitta rätt ord och att sätta ihop ord till bra texter, tydlig information och säljande budskap.',
-        vmid: 'og:description',
+        property: "og:description",
+        content: "Jag gillar att vränga ord, hitta rätt ord och att sätta ihop ord till bra texter, tydlig information och säljande budskap.",
+        vmid: "og:description",
       },
       {
-        name: 'description',
-        content: 'Jag gillar att vränga ord, hitta rätt ord och att sätta ihop ord till bra texter, tydlig information och säljande budskap.',
+        name: "description",
+        content: "Jag gillar att vränga ord, hitta rätt ord och att sätta ihop ord till bra texter, tydlig information och säljande budskap.",
       },
       {
-        property: 'og:type',
-        content: 'website',
-        vmid: 'og:type',
+        property: "og:type",
+        content: "website",
+        vmid: "og:type",
       },
     ],
   },
