+++ src/components/Main.vue
@@ -38,12 +38,12 @@
       <img
         :src="require('@/assets/images/pelle.jpg')"
         class="p-3 lg:p-4 self-center md:justify-self-center who__image block lg:hidden"
-        alt="pelle"
+        alt="Pelle Öhrn"
       />
       <img
         :src="require('@/assets/images/pelle-slanted.png')"
         class="p-3 lg:p-4 self-center who__image hidden lg:block"
-        alt="pelle"
+        alt="Pelle Öhrn"
       />
     </div>
     <div
@@ -107,102 +107,7 @@
         </p>
       </div>
 
-      <form
-        id="js-contact-form"
-        class="contact__form w-full flex flex-col items-center mt-8 lg:mt-0"
-        method="POST"
-      >
-        <div class="flex flex-wrap -mx-3 mb-6 w-full">
-          <div class="w-full md:w-1/2 mb-6 md:mb-0 md:pr-2 lg:pr-3">
-            <label
-              class="block tracking-wide font-body text-darkcyan text-md font-bold mb-2"
-              for="name"
-            >
-              Namn
-            </label>
-            <input
-              id="name"
-              name="name"
-              :class="[
-                'font-body appearance-none block w-full text-darkcyan border-2',
-                'border-gray-300 rounded py-3 px-4 mb-3 leading-tight focus:outline-none',
-                'focus:border-turquoise transition',
-              ]"
-              type="text"
-              placeholder="Namn Namnsson"
-              required
-            />
-          </div>
-          <div class="w-full md:w-1/2 md:pl-2 lg:pl-3">
-            <label
-              class="block tracking-wide font-body text-darkcyan text-md font-bold mb-2"
-              for="email"
-            >
-              Email
-            </label>
-            <input
-              id="email"
-              name="_replyto"
-              :class="[
-                'font-body appearance-none block w-full text-darkcyan border-2',
-                'border-gray-300 rounded py-3 px-4 mb-3 leading-tight focus:outline-none',
-                'focus:border-turquoise transition',
-              ]"
-              type="email"
-              placeholder="exempel@email.se"
-              required
-            />
-          </div>
-        </div>
-        <div class="flex flex-wrap -mx-3 mb-6 w-full">
-          <div class="w-full">
-            <label
-              class="block tracking-wide font-body text-darkcyan text-md font-bold mb-2"
-              for="message"
-            >
-              Meddelande
-            </label>
-            <textarea
-              id="message"
-              name="message"
-              :class="[
-                'font-body appearance-none block w-full text-darkcyan border-2',
-                'border-gray-300 rounded py-3 px-4 mb-3 leading-tight focus:outline-none',
-                'focus:border-turquoise transition h-40 lg:h-auto',
-              ]"
-              type="password"
-              placeholder="Hej! Ses över en kopp kaffe?"
-              required
-            />
-          </div>
-        </div>
-        <input
-          class="hidden"
-          type="text"
-          name="_gotcha"
-          style="display: none"
-        />
-        <input
-          class="hidden"
-          type="hidden"
-          name="_subject"
-          value="Message via http://pelleohrn.se"
-        />
-        <button
-          :class="[
-            'bg-turquoise shadow-md hover:bg-cyan text-darkcyan font-bold font-body',
-            'py-2 px-4 rounded inline-flex items-center transition',
-          ]"
-          type="submit"
-          value="Send"
-        >
-          <font-awesome-icon
-            icon="paper-plane"
-            class="mr-2"
-          />
-          <span>Skicka</span>
-        </button>
-      </form>
+      <Form class="contact__form" />
     </div>
     <div
       class="bg-gray-100 flex flex-col items-center"
@@ -210,29 +115,10 @@
         backgroundImage: 'url(' + require('@/assets/patterns/cubes.png') + ')',
       }"
     >
-      <div class="w-48 relative mt-2 h-40">
-        <div
-          :class="[
-            'w-12 h-12 border-2 border-solid border-gray-600 absolute',
-            'z-0 top-4 left-34 rotate-45 m-auto left-0 right-0',
-          ]"
-        />
-        <div
-          :class="[
-            'shadow-lg absolute w-full mt-6 border-8 border-solid',
-            'bg-white border-frame h-full',
-          ]"
-        >
-          <img
-            src="@/assets/images/painting-by-copywriter.png"
-            class="p-3 h-full shadow-md"
-            alt="painting-by-copywriter"
-          />
-        </div>
-      </div>
+      <Painting />
       <p
         :class="[
-          'font-body text-sm md:text-base text-gray-600 italic w-4/5 lg:w-2/5',
+          'font-body text-md md:text-base text-gray-600 italic w-4/5 lg:w-2/5',
           'max-w-5xl m-auto text-center pt-16',
         ]"
       >
@@ -256,7 +142,14 @@
 </template>
 
 <script>
+import Form from "./Form.vue";
+import Painting from "./Painting.vue";
+
 export default {
+  components: {
+    Form,
+    Painting,
+  },
   data() {
     return {
       email: `${"skri"}${"v@"}pelleohrn.se`,
@@ -266,11 +159,9 @@ export default {
   mounted() {
     /* Delayed to hide email and phone number in prerendering */
     setTimeout(() => {
-      const contactForm = document.getElementById("js-contact-form");
-      contactForm.setAttribute("action", `//formspree.io/${this.email}`);
       document.getElementById("js-contact-phone").innerHTML = this.email;
       document.getElementById("js-contact-email").innerHTML = this.phone;
-    }, 500);
+    }, 1000);
   },
 };
 </script>
@@ -285,21 +176,6 @@ export default {
 .bg-gradient-lightpurple-to-pink {
   background: #fde9ff;
   background: linear-gradient(137deg, #fde9ff 0%, #ffe0f8 100%);
-}
-
-.transition {
-  transition: all 0.2s;
-}
-
-.border-frame {
-  border-top-color: #b89675;
-  border-right-color: #7b654f;
-  border-bottom-color: #7b654f;
-  border-left-color: #a67c52;
-}
-
-.rotate-45 {
-  transform: rotate(45deg);
 }
 
 /* BEM for some grid stuff */
